backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
paramiko==3.5.1
netmiko==4.4.0
napalm==5.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==45.0.6
reportlab==4.4.3
email-validator==2.2.0
